Add files via upload
The previously created 4-bit adder/subtractor has been modified by adding two parallel in, and parallel out registers. These registers are clocked and always have the output enabled. The input enable is given as a toggle, and decides when the inputs of the registers show on the adder/subtractor. The purpose of these registers is to hold the data that is required to be sent to the ALU.