Add files via upload
We would like to store immediate values in the registers RA and RB. We would also like to store ALU results in the same registers. For this specific purpose we have added a multiplexer to the input of the registers to choose between custom input or ALU output to be fed to the registers. In the current design it may seem that the registers receive the same data, however that is because we have used a common enable for RA and RB which will be handled later on during the course of project design. We have also added a D flip flop to store the value of Cout from the ALU to use the JUMP IF CARRY SET command.