Add files via upload
This is a minor iteration over the last version. We have added an output register RO to store the final results once calculations are complete. RO is then used to output the final results! We have added a decoder to select the register to be enabled. RA (00), RB (01), RO (10), NOP (11).